Back up coding

+++ Mod/src/com/mod/interfaces/KiteStockConverter.java
@@ -9,9 +9,9 @@ import com.mod.support.ConfigData;
 
 public class KiteStockConverter {
 	
-	public static final Map<Double, String> KITE_STOCK_LIST = new HashMap<Double, String>();
-	public static final Map<Double, String> BN_PE_LIST = new HashMap<Double, String>();
-	public static final Map<Double, String> BN_CE_LIST = new HashMap<Double, String>();
+	public static final Map<Long, String> KITE_STOCK_LIST = new HashMap<Long, String>();
+	public static final Map<Long, String> BN_PE_LIST = new HashMap<Long, String>();
+	public static final Map<Long, String> BN_CE_LIST = new HashMap<Long, String>();
 	
 	
 	private static ConfigData appConfig(){
@@ -31,12 +31,12 @@ public class KiteStockConverter {
 		String[] values = null;
 		for(int i=0;i<size;i++){
 			values = values(data.get(i));
-			KITE_STOCK_LIST.put(Double.valueOf(values[0]), values[1]);
+			KITE_STOCK_LIST.put(Long.valueOf(values[0]), values[1]);
 			if(values[1].contains("BANKNIFTY") && values[1].contains("CE")){
-				BN_CE_LIST.put(Double.valueOf(values[0]), values[1]);
+				BN_CE_LIST.put(Long.valueOf(values[0]), values[1]);
 			}
 			if(values[1].contains("BANKNIFTY") && values[1].contains("PE")){
-				BN_PE_LIST.put(Double.valueOf(values[0]), values[1]);
+				BN_PE_LIST.put(Long.valueOf(values[0]), values[1]);
 			}
 		}
 	}
