+++ Mod/src/com/mod/interfaces/WebSocketClientExample.java
@@ -9,15 +9,15 @@ import java.security.NoSuchAlgorithmException;
 import java.util.concurrent.CountDownLatch;
 
 import javax.net.ssl.SSLContext;
-import javax.websocket.ClientEndpoint;
-import javax.websocket.CloseReason;
-import javax.websocket.ContainerProvider;
-import javax.websocket.OnClose;
-import javax.websocket.OnError;
-import javax.websocket.OnMessage;
-import javax.websocket.OnOpen;
-import javax.websocket.Session;
-import javax.websocket.WebSocketContainer;
+import jakarta.websocket.ClientEndpoint;
+import jakarta.websocket.CloseReason;
+import jakarta.websocket.ContainerProvider;
+import jakarta.websocket.OnClose;
+import jakarta.websocket.OnError;
+import jakarta.websocket.OnMessage;
+import jakarta.websocket.OnOpen;
+import jakarta.websocket.Session;
+import jakarta.websocket.WebSocketContainer;
 
 import com.mod.datafeeder.DataFeed;
 import com.mod.process.models.CacheService;
@@ -189,7 +189,7 @@ public class WebSocketClientExample {
 		
 		
 		
-		CacheService.PRICE_LIST.put(Double.valueOf(fullObject.getInstrumentToken()), fullObject.getLtp().doubleValue());
+		CacheService.PRICE_LIST.put(Long.valueOf(fullObject.getInstrumentToken()), fullObject.getLtp().doubleValue());
 		/**
 		 * Need meta data in place before it is triggered.
 		 */
